refactor(codegen): extrai calls/ns_call.rs
Move 8 fns de chamadas de namespace ABI (~477 LOC) para
expressions/calls/ns_call.rs:
- emit_constant_load + lower_intrinsic (sqrt/abs/min/max inline)
- lower_ns_call + lower_ns_call_member + lower_ns_call_body
- lower_node_ns_call (redirect Node imports)
- lower_global_instance_call (console.log, Date.now, etc)
- emit_namespace_constant (pub(crate) re-export)

calls/mod.rs: 2994 -> 2521 linhas. Suite verde.
